remove non-template processDiag
