+++ src/io/DiagPlotfile.H
@@ -35,13 +35,6 @@ class DiagPlotfile : public DiagBase::Register<DiagPlotfile>
 	void prepare(int a_nlevels, const amrex::Vector<amrex::Geometry> &a_geoms, const amrex::Vector<amrex::BoxArray> &a_grids,
 		     const amrex::Vector<amrex::DistributionMapping> &a_dmap, const amrex::Vector<std::string> &a_varNames) override;
 
-	// Base class virtual function - not used, template version below is called instead
-	void processDiag(int /*a_nstep*/, const amrex::Real & /*a_time*/, const ProjectionData * /*projectionData*/ = nullptr,
-			 amrex::Direction /*projectionDir*/ = amrex::Direction::x) override
-	{
-		amrex::Abort("DiagPlotfile::processDiag() non-template version should not be called. Use template version instead.");
-	}
-
 	void addVars(amrex::Vector<std::string> &a_varList) override;
 
 	void close() override {}
